Improve the sizing of the image.
We no longer scale the picture so it fits in a predefine size.
We use a default 'scale' (distance between points) and set the
size accordingly. We've also introduced arguments to 'trapped'
for setting margins and the scale. The margins are relative to
the scale (defaults to 1, so the same distance as between to
neighbouring points.